Inicjalny sklep + bind-mount

+++ prestashop-docker-project-1/prestashop/classes/stock/WarehouseProductLocation.php
@@ -24,7 +24,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 /**
- * @deprecated since 9.0 and will be removed in 10.0
+ * @since 1.5.0
  */
 class WarehouseProductLocationCore extends ObjectModel
 {
@@ -82,7 +82,7 @@ class WarehouseProductLocationCore extends ObjectModel
      * @param int $id_product_attribute product attribute ID
      * @param int $id_warehouse warehouse ID
      *
-     * @return string|false $location Location of the product
+     * @return string $location Location of the product
      */
     public static function getProductLocation($id_product, $id_product_attribute, $id_warehouse)
     {
@@ -104,7 +104,7 @@ class WarehouseProductLocationCore extends ObjectModel
      *
      * @param int $id_product
      * @param int $id_product_attribute
-     * @param int $id_warehouse
+     * @param int $id_supplier
      *
      * @return int $id_warehouse_product_location ID of the WarehouseProductLocation
      */
@@ -120,7 +120,7 @@ class WarehouseProductLocationCore extends ObjectModel
 			AND wpl.id_warehouse = ' . (int) $id_warehouse
         );
 
-        return (int) Db::getInstance(_PS_USE_SQL_SLAVE_)->getValue($query);
+        return Db::getInstance(_PS_USE_SQL_SLAVE_)->getValue($query);
     }
 
     /**
